requirements.txt
accelerate==1.12.0
bitsandbytes==0.49.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
filelock==3.20.0
fsspec==2025.12.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
packaging==25.0
pillow==11.3.0
psutil==7.1.3
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1+cu130
torchvision==0.24.1+cu130
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
urllib3==2.6.2
fastapi
pydantic
fastapi[standard]
matplotlib
gradio
sqlalchemy
panda
openpyxl